requirements.txt
h5py~=2.10.0
imageio~=2.9.0
matplotlib~=3.3.4
numpy~=1.19.2
pillow~=8.1.2
scikit-image~=0.17.2
scipy~=1.6.2
# pytorch~=1.8.1
torchvision~=0.9.1
opencv-python~=4.5.1.48

